Rename `StatementPrinter` to `PlainTextStatementFormatter` to more accurately reflect what it is producing
