+++ Theatrical-Players-Refactoring-Kata/Theatrical-Players-Refactoring-KataTests/GenreAmountProviderTests.swift
